// sourcecode/BTrak.Packages/snova-admin-module/admin-module-components/admin-components/src/lib/admin-module/models/customTagsModel.ts
export class CustomTagModel {
    tag: string;
    tags: string;
    referenceId: string;
    searchText: string;
    tagId: string;
    tagsList: any[];
}

export class CustomTagsModel {
    customApplicationId: string;
    genericFormId: string;
    genericFormKeyValue: string;
    tagValue: string;
    genericFormName: string;
    genericFormKeyName: string;
    searchText: string;
    formName: string;
    searchTagText: string;
    genericFormLabel: string;
    genericFormKeyId: string;
}